requirements.txt
# ==============================================================================
# Movie Recommendation System Requirements
# Python 3.10 + CUDA 11.2 Compatible
# ==============================================================================

# ------------------------------------------------------------------------------
# Core Data Processing (STABLE VERSIONS)
# ------------------------------------------------------------------------------
numpy==1.24.3                # Stable, no numpy._core issues
pandas==2.0.3                # Compatible with NumPy 1.24
scikit-learn==1.3.0          # Stable with NumPy 1.24
scipy==1.11.1                # Compatible

# ------------------------------------------------------------------------------
# Visualization
# ------------------------------------------------------------------------------
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# ------------------------------------------------------------------------------
# Web Framework
# ------------------------------------------------------------------------------
flask==2.3.2
werkzeug==2.3.6

# ------------------------------------------------------------------------------
# Deep Learning - TensorFlow (CUDA 11.2)
# ------------------------------------------------------------------------------
tensorflow==2.10.1           # Supports CUDA 11.2 + cuDNN 8.1

# ------------------------------------------------------------------------------
# Deep Learning - PyTorch (CUDA 11.2)
# NOTE: Install separately after other packages:
# pip install torch==1.12.1+cu113 torchvision==0.13.1+cu113 torchaudio==0.12.1+cu113 --index-url https://download.pytorch.org/whl/cu113
# ------------------------------------------------------------------------------
# torch==1.12.1+cu113
# torchvision==0.13.1+cu113
# torchaudio==0.12.1+cu113

# ------------------------------------------------------------------------------
# Configuration & Utilities
# ------------------------------------------------------------------------------
pyyaml==6.0.1
joblib==1.3.1
tqdm==4.65.0

# ------------------------------------------------------------------------------
# Development & Testing
# ------------------------------------------------------------------------------
jupyter==1.0.0
pytest==7.4.0
ipython==8.14.0
notebook==6.5.4

# ------------------------------------------------------------------------------
# Additional Utilities
# ------------------------------------------------------------------------------
python-dotenv==1.0.0
requests==2.31.0
pillow==10.0.0